healthecho-backend/app/modules/pdf_parser.py
from pathlib import Path

import fitz


def parse_pdf_text(file_path: str | Path) -> str:
    path = Path(file_path)
    if not path.exists():
        return ""

    text_parts: list[str] = []
    with fitz.open(path) as doc:
        for page in doc:
            text_parts.append(page.get_text("text"))
    return "\n".join(text_parts).strip()


def parse_pdf_bytes(file_bytes: bytes) -> str:
    text_parts: list[str] = []
    with fitz.open(stream=file_bytes, filetype="pdf") as doc:
        for page in doc:
            text_parts.append(page.get_text("text"))
    return "\n".join(text_parts).strip()
