healthecho-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.2
faiss-cpu==1.9.0
sentence-transformers==3.2.1
pymupdf==1.24.10
pytesseract==0.3.13
pillow==10.4.0
python-dotenv==1.0.1
httpx==0.27.2
numpy==1.26.4
pydantic-settings==2.5.2
